Split IBDO to prevent Extracted Records from having Extracted Records.
